+++ resources/views/home.blade.php
@@ -39,7 +39,7 @@
                                         <td>{{$score->score}}</td>
                                         <td>{{$score->created_at->diffForHumans()}}</td>
                                         <td>
-                                        {!! Form::open(['method'=>'DELETE', 'action'=>['HomeController@destroy', $score->id]]) !!}
+                                        {!! Form::open(['method'=>'DELETE', 'action'=>['HomeController@destroy', $score->id], 'onsubmit' => 'return ConfirmDelete()']) !!}
 
                                         <div class="form-group">
                                         {!!  Form::submit('Delete', ['class'=>'btn btn-danger ']) !!}
@@ -51,26 +51,19 @@
                                 @endforeach
                             @endif
 
-                            {{--@if($userInfo)--}}
-                                {{--@foreach($userInfo as $user)--}}
-                                    {{--<tr>--}}
-                                        {{--<td><img height="50" src="{{$user->photo ? URL::to($user->photo->file) : 'http://localhost/my-archery-pal/public/images/placeholder2.png'}}" ></td>--}}
-{{--                                        <td>{{$user->score->round ? $user->score->round->name : 'No Round Selected'}}</td>--}}
-                                        {{--<td>{{$user->average}}</td>--}}
-{{--                                        <td>{{$user->name}}</td>--}}
-                                        {{--<td>{{$user->created_at->diffForHumans()}}</td>--}}
-                                        {{--<td>--}}
-                                            {{--{!! Form::open(['method'=>'DELETE', 'action'=>['HomeController@destroy', $user->id]]) !!}--}}
+                            <script>
 
-                                            {{--<div class="form-group">--}}
-                                                {{--{!!  Form::submit('Delete', ['class'=>'btn btn-danger ']) !!}--}}
-                                            {{--</div>--}}
+                                function ConfirmDelete()
+                                {
+                                    var x = confirm("Are you sure you want to delete?");
+                                    if (x)
+                                        return true;
+                                    else
+                                        return false;
+                                }
 
-                                            {{--{!!  Form::close() !!}--}}
-                                        {{--</td>--}}
-                                    {{--</tr>--}}
-                                {{--@endforeach--}}
-                            {{--@endif--}}
+                            </script>
+
 
                             </tbody>
                         </table>
